BoardIndex.swift: Created a BoardIndex class that will hold the indices data for the chess pieces, with a static func to return the 1D array value of a given input boardIndex.
ChessPiece.swift
Created a ChessPiece class that holds the minimal common properties of all the chessPieces regardless of their color and/or type.

Rook.swift
Created a Rook class that holds the info of the rook, init takes a color from ChessPieceColor enum and setup the parameters (image, rank, index) of the super.init according to the color given as input.

Themes.swift:
Created a ChessPieceColor enum that holds the different colors/themes for the chess pieces, allows for further addition of different themes in the future. (Roman, Egyptian, ..)

ViewController.swift :
Created a blackArray of ChessPiece that will hold the different ChessPiece subclasses of black color.
Created a whiteArray of ChessPiece that will hold the different ChessPiece subclasses of white color.
setupInitialBoard() function that will place the pieces in their initial position when the game starts, and add the piece to it's proper array. Only rooks are being created at this point, switch cases are grouped for common statements.