Created multiple functions to help color the board for the user to provide assist on possible movements.
Added check valid move for the ChessPiece class that will be override by the subclasses according to their movement logic.

Added an ID property to the ChessPiece to check if it's the same piece or not, also a static counter that will be the id to each piece.

Added coloring properties to further improve the ability to add more themes in the future.

Added boardPieceArray that will hold all the pieces.

Added updateUI function that will reflect the boardPiecesArray model on the buttonArray (view).