Massive Commit with massive changes, defo worth it.
Create a ChessBoard that is responsible for the board operations/data.

ChessPiece now supports the equatable protocol.

Added the right rook logic, same row or same col.

Massive changes to the themes as now the different themes are properly supported. The additional themes will be added only once and set once the game has started.

Too much changes in the controller to list all of them but there was a great overhaul of the logic used there.